+++ cpu/include/solver_red_black.h
@@ -9,7 +9,7 @@
 
 class SolverRedBlack : public Solver {
 public:
-    SolverRedBlack(double* grid, int w, int h, const std::string& name);
+    SolverRedBlack(float* grid, int w, int h, const std::string& name);
     ~SolverRedBlack();
     // MODIFIED SIGNATURE:
     void solve(const SimulationParameters& sim_params) override;
