+++ cuda/tests/test_solver_basic.cpp
@@ -46,10 +46,10 @@ std::string getProjectDir() {
 }
 
 /* ---------- simple CSV reader --------------------------------- */
-static std::vector<std::vector<double>>
+static std::vector<std::vector<float>>
 read_csv(const std::string& file, int w, int h)
 {
-    std::vector<std::vector<double>> grid(h, std::vector<double>(w, 0.0));
+    std::vector<std::vector<float>> grid(h, std::vector<float>(w, 0.0));
     std::ifstream in(file);
     std::string line;
     int y = 0;
@@ -57,11 +57,11 @@ read_csv(const std::string& file, int w, int h)
         size_t start = 0, end = line.find(',');
         int x = 0;
         while (end != std::string::npos && x < w) {
-            grid[y][x++] = std::stod(line.substr(start, end - start));
+            grid[y][x++] = static_cast<float>(std::stod(line.substr(start, end - start)));
             start = end + 1;
             end   = line.find(',', start);
         }
-        if (x < w && start < line.size()) grid[y][x] = std::stod(line.substr(start));
+        if (x < w && start < line.size()) grid[y][x] = static_cast<float>(std::stod(line.substr(start)));
         ++y;
     }
     return grid;
@@ -74,24 +74,24 @@ int main() {
     const int height = 50;
 
     /* ---- boundary & host grid -------------------------------- */
-    BoundaryConditions bc{0.0, 0.0, 0.0, 0.0};
-    std::vector<double> U_host(width * height, 0.0);
+    BoundaryConditions bc{0.0f, 0.0f, 0.0f, 0.0f};
+    std::vector<float> U_host(width * height, 0.0f);
     initializeGrid(U_host.data(), width, height, bc);
 
     /* ---- device buffer -------------------------------------- */
-    double* d_U = nullptr;
-    CUDA_CHECK_ERROR(cudaMalloc(&d_U, width * height * sizeof(double)));
+    float* d_U = nullptr;
+    CUDA_CHECK_ERROR(cudaMalloc(&d_U, width * height * sizeof(float)));
     CUDA_CHECK_ERROR(cudaMemcpy(d_U, U_host.data(),
-                                width * height * sizeof(double),
+                                width * height * sizeof(float),
                                 cudaMemcpyHostToDevice));
 
     /* ---- solver --------------------------------------------- */
     SolverBasic solver(d_U, width, height, "SolverBasic_Test");
-    SimulationParameters params{width, height, 1e-5, 20000, 1.9};   // NEW
+    SimulationParameters params{width, height, 1e-5f, 20000, 1.9f};   // NEW
     solver.solve(params);                                           // NEW
 
     CUDA_CHECK_ERROR(cudaMemcpy(U_host.data(), d_U,
-                                width * height * sizeof(double),
+                                width * height * sizeof(float),
                                 cudaMemcpyDeviceToHost));
 
     /* ---- export & validation -------------------------------- */
