+++ include/grid_initialization.h
@@ -4,6 +4,6 @@
 
 #include "boundary_conditions.h" // Include the header where BoundaryConditions is defined
 
-void initializeGrid(double *U, int width, int height, const BoundaryConditions& bc);
+void initializeGrid(float *U, int width, int height, const BoundaryConditions& bc);
 
 #endif // GRID_INITIALIZATION_H
